Remove unused code and most circular dependencies
