+++ packages/node/src/node.ts
@@ -3,7 +3,6 @@ import { BaseProvider } from "ethers/providers";
 import { SigningKey } from "ethers/utils";
 import { fromExtendedKey, HDNode } from "ethers/utils/hdnode";
 import EventEmitter from "eventemitter3";
-import log from "loglevel";
 import { Memoize } from "typescript-memoize";
 
 import { createRpcRouter } from "./api";
@@ -20,7 +19,7 @@ import { RequestHandler } from "./request-handler";
 import RpcRouter from "./rpc-router";
 import { getHDNode } from "./signer";
 import { NODE_EVENTS, NodeMessageWrappedProtocolMessage } from "./types";
-import { timeout } from "./utils";
+import { debugLog, timeout } from "./utils";
 
 export interface NodeConfig {
   // The prefix for any keys used in the store by this Node depends on the
@@ -92,15 +91,15 @@ export class Node {
 
     this.instructionExecutor = this.buildInstructionExecutor();
 
-    log.info(
+    debugLog(
       `Waiting for ${this.blocksNeededForConfirmation} block confirmations`
     );
   }
 
   private async asynchronouslySetupUsingRemoteServices(): Promise<Node> {
     this.signer = await getHDNode(this.storeService);
-    log.info(`Node signer address: ${this.signer.address}`);
-    log.info(`Node public identifier: ${this.publicIdentifier}`);
+    debugLog(`Node signer address: ${this.signer.address}`);
+    debugLog(`Node public identifier: ${this.publicIdentifier}`);
     this.requestHandler = new RequestHandler(
       this.publicIdentifier,
       this.incoming,
@@ -383,31 +382,4 @@ export class Node {
  */
 export function getETHFreeBalanceAddress(publicIdentifier: string) {
   return fromExtendedKey(publicIdentifier).derivePath("0").address;
-}
-
-const isBrowser =
-  typeof window !== "undefined" &&
-  {}.toString.call(window) === "[object Window]";
-
-export function debugLog(...messages: any[]) {
-  try {
-    const logPrefix = "NodeDebugLog";
-    if (isBrowser) {
-      if (localStorage.getItem("LOG_LEVEL") === "DEBUG") {
-        // for some reason `debug` doesn't actually log in the browser
-        log.info(logPrefix, messages);
-        log.trace();
-      }
-      // node.js side
-    } else if (
-      process.env.LOG_LEVEL !== undefined &&
-      process.env.LOG_LEVEL === "DEBUG"
-    ) {
-      log.debug(logPrefix, JSON.stringify(messages, null, 4));
-      log.trace();
-      log.debug("\n");
-    }
-  } catch (e) {
-    console.error("Failed to log: ", e);
-  }
 }
