+++ packages/node/src/protocol/uninstall.ts
@@ -1,10 +1,10 @@
 import { BaseProvider } from "ethers/providers";
 
 import { SetStateCommitment } from "../ethereum";
-import { Protocol, ProtocolExecutionFlow } from "../machine";
-import { Opcode } from "../machine/enums";
+import { Opcode, Protocol } from "../machine/enums";
 import {
   Context,
+  ProtocolExecutionFlow,
   ProtocolMessage,
   ProtocolParameters,
   UninstallParams
